Refactor: Improve project structure, CSS, and add tests
This commit addresses several areas identified in the initial project review:

Component Refactoring:
- I separated ServiceCard from ServiceSection, placing them in individual files (`ServiceCard.tsx`, `ServiceSection.tsx`) and removing the duplicate ServiceSection directory.
- I separated TestimonialCard from TestimonialSection, placing them in individual files (`TestimonialCard.tsx`, `TestimonialSection.tsx`).
- I updated component exports in `src/components/index.ts` accordingly.

CSS Consolidation:
- I removed duplicate CSS rules for `.card`, `.btn`, `.layout-container`, and `.section-spacing` from `App.css`.
- I prioritized Tailwind CSS utilities and `index.css` as the source of truth for these styles.

Core Functionality & Placeholders:
- I implemented basic search functionality in the Hero component (state management and console log on search).
- I changed Navbar "Get Started" buttons from `<button>` to `<a>` tags for semantic navigation.
- I updated placeholder `href` attributes in the Footer for company, legal, and social media links.
- I reviewed TestimonialCard image handling, confirmed fallback logic, and added a TODO for future enhancements.

Configuration & Dependencies (Review Only):
- I manually reviewed versions of key dependencies (`@mui/material`, `react`, `typescript`) and noted areas for future attention.
- I manually reviewed ESLint (modern flat config, but missing accessibility plugin) and PostCSS (experimental stage 1, potential plugin redundancy) configurations.

Testing:
- I introduced Vitest as the testing framework.
- I installed `@testing-library/react`, `@testing-library/jest-dom`, and `jsdom`.
- I added Vitest configuration (`vitest.config.ts`) and test setup file (`src/setupTests.ts`).
- I included a basic smoke test for the Hero component.
- I updated `tsconfig.app.json` for Vitest globals.
- I added a "test" script to `package.json`.
