+++ src/App.css
@@ -5,70 +5,3 @@
   padding: 0;
   text-align: left;
 }
-
-/* Layout containers */
-.container {
-  width: 100%;
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 0 1rem;
-}
-
-/* Section spacing */
-section {
-  padding: 4rem 0;
-}
-
-/* Card styles */
-.card {
-  background: white;
-  border-radius: 1rem;
-  padding: 2rem;
-  transition: all 0.3s ease;
-}
-
-.card:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.1);
-}
-
-/* Button styles */
-.btn {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0.75rem 1.5rem;
-  border-radius: 9999px;
-  font-weight: 600;
-  transition: all 0.3s ease;
-}
-
-.btn-primary {
-  background-color: var(--primary);
-  color: white;
-}
-
-.btn-primary:hover {
-  background-color: var(--primary-dark);
-}
-
-.btn-outline {
-  border: 2px solid var(--primary);
-  color: var(--primary);
-}
-
-.btn-outline:hover {
-  background-color: var(--primary);
-  color: white;
-}
-
-/* Standardize padding and margin for layout containers */
-.layout-container {
-  padding: 2rem;
-  margin: 1rem 0;
-}
-
-/* Adjust section spacing to be consistent */
-.section-spacing {
-  padding: 3rem 0;
-}
