+++ src/components/ServiceCard/ServiceSection.tsx
@@ -1,20 +1,4 @@
-interface ServiceCardProps {
-  title: string;
-  description: string;
-  icon: React.ReactNode;
-}
-
-export const ServiceCard = ({ title, description, icon }: ServiceCardProps) => {
-  return (
-    <div className="card hover-scale p-6 rounded-2xl shadow-custom">
-      <div className="w-12 h-12 bg-emerald-100 rounded-xl flex items-center justify-center text-emerald-600 mb-4">
-        {icon}
-      </div>
-      <h3 className="text-xl font-semibold text-gray-900 mb-3">{title}</h3>
-      <p className="text-gray-600 leading-relaxed">{description}</p>
-    </div>
-  );
-};
+import { ServiceCard } from './ServiceCard';
 
 export const ServiceSection = () => {
   const services = [
