+++ src/components/index.ts
@@ -2,8 +2,10 @@
 export * from './Navbar';
 export * from './Footer';
 export * from './Hero';
-export * from './ServiceCard';
-export * from './TestimonialCard';
+export * from './ServiceCard/ServiceCard';
+export * from './ServiceCard/ServiceSection';
+export * from './TestimonialCard/TestimonialCard';
+export * from './TestimonialCard/TestimonialSection';
 export * from './HowItWorks';
 export * from './TrustSignals';
 export * from './CallToAction';
